- Refactoring of features package
