+++ feature/home/src/main/java/com/exalt/feature/home/viewobjects/PostVO.kt
@@ -1,0 +1,11 @@
+package com.exalt.feature.home.viewobjects
+
+data class PostVO (
+    val id: String,
+    val text: String,
+    val imageUri: String,
+    val publishDate: String,
+    val ownerId: String,
+    val ownerName: String,
+    val ownerPictureUri: String
+)
