+++ feature/home/src/main/res/navigation/home_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.exalt.home.fragments.HomeFragment"
+        android:name="com.exalt.feature.home.fragments.HomeFragment"
         android:label="HomeFragment"
         tools:layout="@layout/fragment_home"/>
 </navigation>
