+++ feature/post/src/main/java/com/exalt/feature/post/viewobjects/CommentVO.kt
@@ -1,0 +1,10 @@
+package com.exalt.feature.post.viewobjects
+
+data class CommentVO(
+    val id: String,
+    val message: String,
+    val publishDate: String,
+    val ownerId: String,
+    val ownerName: String,
+    val ownerPictureUri: String
+)
