+++ feature/post/src/main/java/com/exalt/feature/post/viewobjects/PostVO.kt
@@ -1,0 +1,13 @@
+package com.exalt.feature.post.viewobjects
+
+data class PostVO (
+    val id: String,
+    val text: String,
+    val imageUri: String,
+    val publishDate: String,
+    val ownerId: String,
+    val ownerName: String,
+    val ownerPictureUri: String,
+    val tags: List<String>,
+    val likes: Int
+)
