--- feature/post/src/main/java/com/exalt/post/viewmodels/PostViewModel.kt
@@ -1,44 +1,0 @@
-package com.exalt.post.viewmodels
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.liveData
-import com.exalt.domain.home.usecases.GetCommentsByPostUseCase
-import com.exalt.domain.home.usecases.GetPostUseCase
-import com.exalt.post.mappers.CommentVoMapper
-import com.exalt.post.mappers.PostVoMapper
-import com.exalt.post.viewobjects.CommentVO
-import com.exalt.post.viewobjects.PostVO
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class PostViewModel @Inject constructor(
-    private val getPostUseCase: GetPostUseCase,
-    private val getCommentsUseCase: GetCommentsByPostUseCase,
-    private val commentVoMapper: CommentVoMapper,
-    private val postVoMapper: PostVoMapper,
-
-    ): ViewModel() {
-
-    private val _isLoading = MutableLiveData(true)
-    val isLoading: LiveData<Boolean> = _isLoading
-
-    var postId = ""
-    private var commentsPage = 0u
-
-    val post: LiveData<PostVO> = liveData {
-        emit(
-            postVoMapper.toPostVO(getPostUseCase.invoke(postId))
-        )
-    }
-
-    val comments: LiveData<List<CommentVO>> = liveData {
-        emit(
-            commentVoMapper.toListCommentVO(getCommentsUseCase.invoke(postId, commentsPage))
-        )
-        _isLoading.value = false
-    }
-
-}
