--- feature/post/src/main/java/com/exalt/post/viewobjects/CommentVO.kt
@@ -1,10 +1,0 @@
-package com.exalt.post.viewobjects
-
-data class CommentVO(
-    val id: String,
-    val message: String,
-    val publishDate: String,
-    val ownerId: String,
-    val ownerName: String,
-    val ownerPictureUri: String
-)
