--- feature/post/src/main/java/com/exalt/post/viewobjects/PostVO.kt
@@ -1,13 +1,0 @@
-package com.exalt.post.viewobjects
-
-data class PostVO (
-    val id: String,
-    val text: String,
-    val imageUri: String,
-    val publishDate: String,
-    val ownerId: String,
-    val ownerName: String,
-    val ownerPictureUri: String,
-    val tags: List<String>,
-    val likes: Int
-)
