+++ feature/post/src/main/res/layout/fragment_post.xml
@@ -32,7 +32,6 @@
             android:layout_height="wrap_content"
             android:layout_weight="1">
 
-
             <ImageView
                 android:id="@+id/post_picture"
                 android:layout_width="match_parent"
