+++ feature/user/src/main/java/com/exalt/feature/user/fragments/UserFragment.kt
@@ -1,0 +1,6 @@
+package com.exalt.feature.user.fragments
+
+import androidx.fragment.app.Fragment
+
+class UserFragment: Fragment() {
+}
