--- feature/user/src/main/java/com/exalt/user/fragments/UserFragment.kt
@@ -1,6 +1,0 @@
-package com.exalt.user.fragments
-
-import androidx.fragment.app.Fragment
-
-class UserFragment: Fragment() {
-}
